Implement ArgTypeMethods and AbiMethods.
* `store_fn_arg` from `ArgTypeMethods` stores a function argument of type
`ArgType` into a specified destination (a `PlaceRef`). This commit handles some
of the possible `PassMode`s of the argument.
* `new_fn_type` from `AbiMethods` returns the ironox function type for a given
`ty::FnSig`. To simplify the implementation (and to be consistent with the
way the LLVM backend achieves this), I defined an `FnTypeExt` trait. This trait
is implemented for `FnType`s. The functions in `FnTypeExt` make it easy to
create a `FnType` from a `ty::FnSig`, and to convert the `FnType` to an ironox
function type.