Return the index of a `Type` stored in `type_cache`.
Before, `add_type` returned `types.len()`, regardless of the input. This was
incorrect, because `add_type` is supposed to return the index of a `Type` in the
`types` vector.

If the `ll_type` is already in `type_cache`, `add_type` must return its index
(retrieved from `type_cache`). Otherwise, the `ll_type` is added to `types` and
`type_cache`, and its index is returned (in this case, the index is
`types.len()`, because the type is added at the end of `types`).