Only run the assembler on regular modules.
There are 3 types of modules: `regular`, `metadata`, and `allocator`.

For `regular` modules (the ones that contain the IR), `no_integrated_as` can be
set by passing `no-integrated-as` to `rustc`. `no_integrated_as` is always false
for `metadata` and `allocator` modules (as far as I know, they don't actually
require an assembler anyway). Still, `codegen` is called for every module,
including `metadata` and `allocator` ones. `run_assembler` should only be run on
`regular` modules for which `no_integrated_as = true` (because `run_assembler`
runs the GNU assembler, as opposed to the 'internal' one). I had to make these
changes, because the compiler was panicking for `metadata` and `allocator`
modules (at `bug!`).